Move tests against static Dogstatsd API into specific test fixture designed to only configure once and run all tests single threaded
